// src/types/OutboundType.ts
export type OutboundType = {
    id: string;
    customer_id: string;
    customer_name: string;
    product_id: string;
    product_name: string;
    unit: string;
    cost_per_unit: number;
    quantity: number;
    price_per_unit: number;
    shipment_date: string;
    total_price: number;
    net_profit: number;
    remark: string;
  };
  